moved contact form to a seperate jsx component and solved metadata issue on use client component
